backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
python-dotenv
Flask-Login
email_validator
